Add files via upload
A mixed strategy improved dung beetle optimization (MSDBO) algorithm is proposed to address the problems of slow convergence speed, easy falling into local optimum, and insufficient search accuracy of the dung beetle optimization algorithm. Firstly, the good point set strategy is introduced to initialize the population and improve the population diversity. Then, the spiral search strategy is combined with the whale optimization algorithm to improve the location update of dung beetle reproduction and foraging behavior, balancing the local exploitation and global search ability of the algorithm, and improving the convergence ability of the algorithm. Finally, the Levy flight strategy is used to improve the location update of dung beetle stealing behavior and improve the algorithm's ability to jump out of the local the ability of the algorithm to jump out of local optimality. The results, tested with 12 benchmark functions and validated using the Wilcoxon rank sum test, show that the improved algorithm has significant advantages in terms of convergence speed, stability, and solution accuracy. In addition, we also applied the MSDBO algorithm to a two-dimensional maximum entropy image segmentation task, and the experimental results show that the MSDBO algorithm has good performance in image segmentation.